+++ mysql_rep_change.py
@@ -402,25 +402,41 @@ def move_slave_up(master, slaves, **kwargs):
         defaults_file=master.defaults_file, rep_user=master.rep_user,
         rep_japd=master.rep_japd)
     slv_master.connect()
-    err_flag, err_msg = mysql_libs.sync_rep_slv(new_master, slv_master)
 
-    if err_flag:
+    if new_master.conn_msg or slv_master.conn_msg:
+        err_flag = True
+        err_msg = "Detected problem in one of the connections"
+
+        print("Error:  Connection problem for new master/slave master.")
+        print("\tNew Master:  %s" % (new_master.conn_msg))
+        print("\tSlave Master:  %s" % (slv_master.conn_msg))
+
+        if new_master.conn:
+            cmds_gen.disconnect(new_master)
+
+        if slv_master.conn:
+            cmds_gen.disconnect(slv_master)
+
+    else:
+        err_flag, err_msg = mysql_libs.sync_rep_slv(new_master, slv_master)
+
+        if err_flag:
+            cmds_gen.disconnect(new_master, slv_master)
+            return err_flag, err_msg
+
+        err_flag, err_msg = \
+            mysql_libs.sync_rep_slv(
+                master, mysql_libs.find_name(slaves, kwargs.get("slv_mv")))
+
+        if err_flag:
+            cmds_gen.disconnect(new_master, slv_master)
+            return err_flag, err_msg
+
+        mysql_libs.change_master_to(new_master, slave_move)
+        mysql_libs.chg_slv_state([slave_move, slv_master], "start")
+        is_slv_up(slave_move)
+        is_slv_up(slv_master)
         cmds_gen.disconnect(new_master, slv_master)
-        return err_flag, err_msg
-
-    err_flag, err_msg = \
-        mysql_libs.sync_rep_slv(
-            master, mysql_libs.find_name(slaves, kwargs.get("slv_mv")))
-
-    if err_flag:
-        cmds_gen.disconnect(new_master, slv_master)
-        return err_flag, err_msg
-
-    mysql_libs.change_master_to(new_master, slave_move)
-    mysql_libs.chg_slv_state([slave_move, slv_master], "start")
-    is_slv_up(slave_move)
-    is_slv_up(slv_master)
-    cmds_gen.disconnect(new_master, slv_master)
 
     return err_flag, err_msg