+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - config/mysql.cfg.TEMPLATE:  Point to correct socket file.
 
 ### Changed
+- move_slave_up:  Changed arguments from positional to keywords for SlaveRep instance call.
 - crt_slv_mst:  Added replication user to MasteRep instance call.
 - config/mysql_cfg.py.TEMPLATE:  Changed entry to work with mysql_libs v5.0.0 library.
 - config/slave.txt.TEMPLATE:  Added rep_user and rep_japd entries to configuration file.