+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - config/mysql.cfg.TEMPLATE:  Point to correct socket file.
 
 ### Changed
-- move_slave_up, run_program:  Replaced cmds_gen.disconnect call with mysql_libs.disconnect call.
+- move_slave, move_slave_up, run_program:  Replaced cmds_gen.disconnect call with mysql_libs.disconnect call.
 - move_slave_up, move_slave:  Replaced fetch_slv call with mysql_libs.fetch_slv call.
 - crt_slv_mst:  Added master connection status checks.
 - move_slave_up:  Added new master and slave master connection status checks.